Bring together ComplianceBanner module
