+++ packages/react-composites/src/composites/CallComposite/components/ComplianceBanner/index.tsx
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+export { ComplianceBanner } from './ComplianceBanner';
+export type { ComplianceBannerProps, ComplianceBannerStrings } from './ComplianceBanner';
